feat(wapod): Make max_instances optional and determined by available parallelism
- `wapod/src/main.rs`:
  - Introduced `available_parallelism` to dynamically determine default parallel instances if `max_instances` is not provided.
  - Changed `max_instances` type to `Option<usize>` in `Args`.
  - Implemented `Args::max_instances` method to handle optional value and fallback to available parallelism.

- `wapod/src/web_api.rs`:
  - Updated to use the `Args::max_instances` method for determining the number of instances.

- `wapod/src/web_api/state.rs`:
  - Modified to use `Args::max_instances` method for consistent instance count handling.

+++ wapod/src/web_api.rs
@@ -303,7 +303,7 @@ fn sign_http_response(data: &[u8]) -> Option<String> {
 pub fn crate_worker_state(args: Args) -> Result<Worker> {
     let (tx, mut rx) = crate_outgoing_request_channel();
     let (run, spawner) = service::service(
-        args.max_instances.saturating_add(2) as usize,
+        args.max_instances().saturating_add(2),
         args.module_cache_size,
         tx,
         &args.blobs_dir,
